rename heading detection module dt file
