+++ server/src/processing/HeadingDetectionDTModule/HeadingDetectionDTModule.ts
